flask/requirements.txt
numpy==1.26.4
torch==2.4.1
sentence-transformers
requests
openai
elasticsearch
flask
python-dotenv
